Add patch for cross-riscv-none-elf
Bare metal RISC-V tool chain requires a patch for successful build with
Void Linux xbps-src build system. This is a work around as such patch is
not required for native tool chain build and it should not be used in
upstream package. Leave it here until a better option will be found.